requirements.txt
flask==3.0.0
twilio==8.10.0
openai==0.28.0
python-dotenv==1.0.0
